0.1: Phase 1 - Functional Maven Generation
This commit is the first phase of the toolchain rewrite project. The second phase is integration into new ForgeGradle. See the README.md for more information (still incomplete).

Please note that the code is still not very clean. However, it is functional and does what it needs to do. We can clean it up over time and there may still be some existing pain points, but the fundemental functionality of MinecraftMaven is now complete and implemented.