Set status attribute to "release" for built modules
If `org.gradle.status` is unspecified in a Gradle module, Gradle will read it as "integration". This is a problem, because in the event of a version conflict, Gradle will always prefer the "release" status.

Consider this example:
- "1.21.11-61.0.1" uses "release"
- "1.21.11-61.0.2" uses "integration"

Gradle's version parser doesn't understand version numbers with multiple versions in them, so it treats the Forge version as a qualifier. In this case, 61.0.1 always wins since it is marked as a "release" version.